+++ plone/app/users/vocabularies.py
@@ -4,14 +4,15 @@ from Products.CMFCore.permissions import ManagePortal
 from Products.CMFCore.utils import getToolByName
 from Products.CMFPlone.utils import normalizeString
 from Products.CMFPlone.utils import safe_unicode
-from plone.app.users.browser.register import RegistrationForm
 from zope.interface import implements
+from zope.component import getUtility
 from zope.schema import getFieldNames
 from zope.schema.interfaces import IVocabularyFactory
 from zope.schema.vocabulary import SimpleVocabulary
 from zope.schema.vocabulary import SimpleTerm
 from zope.site.hooks import getSite
 
+from .schema import IRegisterSchemaProvider
 
 # Define constants from the Join schema that should be added to the
 # vocab of the join fields setting in usergroupssettings controlpanel.
@@ -37,8 +38,7 @@ class UserRegistrationFieldsVocabulary(object):
       >>> len(fields.by_token)
       10
       >>> [k.value for k in fields] # doctest: +NORMALIZE_WHITESPACE
-      ['username', 'description', 'home_page', 'email', 'password_ctl',
-      'portrait', 'fullname', 'password', 'mail_me', 'location']
+      ['username', 'description', 'home_page', 'email', 'password_ctl', 'portrait', 'fullname', 'password', 'mail_me', 'location']
 
       >>> email = fields.by_token['email']
       >>> email.title, email.token, email.value
@@ -49,7 +49,9 @@ class UserRegistrationFieldsVocabulary(object):
 
     def __call__(self, context):
         # default list of Registration Form fields
-        values = getFieldNames(RegistrationForm.schema)
+        util = getUtility(IRegisterSchemaProvider)
+        schema = util.getSchema()
+        values = getFieldNames(schema)
 
         # make sure required minimum number of fields is present
         for val in JOIN_CONST: